Add back trust all certificates
