+++ driver/src/main/java/org/neo4j/driver/Config.java
@@ -34,6 +34,7 @@ import org.neo4j.driver.net.ServerAddressResolver;
 import org.neo4j.driver.util.Immutable;
 import org.neo4j.driver.util.Resource;
 
+import static org.neo4j.driver.Config.TrustStrategy.trustSystemCertificates;
 import static org.neo4j.driver.Logging.javaUtilLogging;
 
 /**
@@ -249,7 +250,7 @@ public class Config
         private long maxConnectionLifetimeMillis = PoolSettings.DEFAULT_MAX_CONNECTION_LIFETIME;
         private long connectionAcquisitionTimeoutMillis = PoolSettings.DEFAULT_CONNECTION_ACQUISITION_TIMEOUT;
         private boolean encrypted = false;
-        private TrustStrategy trustStrategy = TrustStrategy.trustSystemCertificates();
+        private TrustStrategy trustStrategy = trustSystemCertificates();
         private int routingFailureLimit = RoutingSettings.DEFAULT.maxRoutingFailures();
         private long routingRetryDelayMillis = RoutingSettings.DEFAULT.retryTimeoutDelay();
         private long routingTablePurgeDelayMillis = RoutingSettings.DEFAULT.routingTablePurgeDelayMs();
@@ -687,6 +688,7 @@ public class Config
          */
         public enum Strategy
         {
+            TRUST_ALL_CERTIFICATES,
             TRUST_CUSTOM_CA_SIGNED_CERTIFICATES,
             TRUST_SYSTEM_CA_SIGNED_CERTIFICATES
         }
@@ -781,6 +783,17 @@ public class Config
         public static TrustStrategy trustSystemCertificates()
         {
             return new TrustStrategy( Strategy.TRUST_SYSTEM_CA_SIGNED_CERTIFICATES );
+        }
+
+        /**
+         * Trust strategy for certificates that trust all certificates blindly. Suggested to only use this in tests.
+         *
+         * @return an authentication config
+         * @since 1.1
+         */
+        public static TrustStrategy trustAllCertificates()
+        {
+            return new TrustStrategy( Strategy.TRUST_ALL_CERTIFICATES );
         }
     }
 }
