--- driver/src/test/java/org/neo4j/driver/util/SecurityUtil.java
@@ -1,66 +1,0 @@
-/*
- * Copyright (c) 2002-2019 "Neo4j,"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.driver.util;
-
-import java.security.GeneralSecurityException;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
-
-import org.neo4j.driver.internal.security.SecurityPlan;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-public class SecurityUtil
-{
-    public static SecurityPlan trustAllCertificates() throws GeneralSecurityException
-    {
-        SecurityPlan plan = mock( SecurityPlan.class );
-        when( plan.requiresEncryption() ).thenReturn( true );
-        when( plan.requiresHostnameVerification() ).thenReturn( false );
-
-        SSLContext sslContext = SSLContext.getInstance( "TLS" );
-        sslContext.init( new KeyManager[0], new TrustManager[]{new TrustAllTrustManager()}, null );
-        when( plan.sslContext() ).thenReturn( sslContext );
-
-        return plan;
-    }
-
-    private static class TrustAllTrustManager implements X509TrustManager
-    {
-        public void checkClientTrusted( X509Certificate[] chain, String authType ) throws CertificateException
-        {
-            throw new CertificateException( "All client connections to this client are forbidden." );
-        }
-
-        public void checkServerTrusted( X509Certificate[] chain, String authType ) throws CertificateException
-        {
-            // all fine, pass through
-        }
-
-        public X509Certificate[] getAcceptedIssuers()
-        {
-            return new X509Certificate[0];
-        }
-    }
-}
